% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{admin}
\alias{admin}
\title{Paramétrages administrateur}
\usage{
admin(
  commune = "",
  parcelle_ini = "",
  mail = "",
  host = "",
  username_smtp = "",
  password_smtp = "",
  port_smtp = "",
  adresse = "",
  administrateur = "",
  titre_administrateur = ""
)
}
\arguments{
\item{commune}{nom de la commune}

\item{parcelle_ini}{nom de la parcelle à afficher à l'ouverture de la carte (ex: Cheny_00A_0201)}

\item{mail}{adresse mail de l'administrateur}

\item{host}{hôte smtp du mail de l'administrateur}

\item{username_smtp}{compte smtp de l'administrateur}

\item{password_smtp}{mot de passe compte mail administrateur}

\item{port_smtp}{port smtp}

\item{adresse}{adresse http du site de l'application}

\item{administrateur}{nom de l'administrateur}

\item{titre_administrateur}{titre de l'administrateur (ex: 'propriétaire forestier...')}
}
\value{
rien: met à jour le fichier admin.csv
}
\description{
Paramétrages administrateur
}
