% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{admin_backup}
\alias{admin_backup}
\title{Sauvegarde des données}
\usage{
admin_backup()
}
\value{
copie des fichiers
}
\description{
db.sqlite, admin.csv et parcelles.rds sous le dossier indiqué dans path.txt
}
