% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{admin_check}
\alias{admin_check}
\title{Import des données depuis un dossier externe}
\usage{
admin_check()
}
\value{
rien. Remplace admin.csv et le shapefile BDPARCELLE de data par ceux du dossier externe,
et construit parcelles.rds.
}
\description{
Le chemin du dossier externe est indiqué dans le fichier inst/path.txt du package:
à défaut, $HOME/boursefonciereforestiere
}
