% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{affiche_controles}
\alias{affiche_controles}
\title{Montre ou cache les controles selon le contexte}
\usage{
affiche_controles(r, input, output, session)
}
\arguments{
\item{r}{r}

\item{input}{input}

\item{output}{output}

\item{session}{session}
}
\value{
rien
}
\description{
Montre ou cache les controles selon le contexte
}
