% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{bilan_proprietaire}
\alias{bilan_proprietaire}
\title{Table récapitulatives d'un propriétaire}
\usage{
bilan_proprietaire(data, user)
}
\arguments{
\item{data}{issu de r$data}

\item{user}{issu de r$user}
}
\value{
liste: proprietes = tables des parcelles prorpriétées de user ; 
aquisitions = tables des parcelles qui intéressent user;
bilan = bilan des échanges
}
\description{
Table récapitulatives d'un propriétaire
}
