% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{fillcolor}
\alias{fillcolor}
\title{Couleurs de remplissage selon propriétaire}
\usage{
fillcolor(r)
}
\arguments{
\item{r}{r}
}
\value{
vecteur des couleurs des parcelles
}
\description{
Couleurs de remplissage selon propriétaire
}
