% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{get_parcelle}
\alias{get_parcelle}
\title{Récupère les parcelles du groupe d'une parcelle}
\usage{
get_parcelle(r, pc)
}
\arguments{
\item{r}{r}

\item{pc}{nom de la parcelle}
}
\value{
vecteur des parcelles du même groupe que pc
}
\description{
Récupère les parcelles du groupe d'une parcelle
}
