% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{ico}
\alias{ico}
\title{Icones des parcelles}
\usage{
ico(r)
}
\arguments{
\item{r}{r}
}
\value{
liste: centro = sf des centroïdes des parcelles, ico = vecteur des nom des icones
}
\description{
Icones des parcelles
}
