% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{init_table_identite}
\alias{init_table_identite}
\title{Réinitialisation de la table identite}
\usage{
init_table_identite()
}
\value{
rien: écrit une sauvegarde de l'ancienne table au format rds, et écrit une nouvelle table vide.
}
\description{
Réinitialisation de la table identite
}
