% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{init_table_interet}
\alias{init_table_interet}
\title{Réinitialisation de la table interet}
\usage{
init_table_interet()
}
\value{
rien: écrit une sauvegarde de l'ancienne table au format rds, et écrit une nouvelle table vide.
}
\description{
Réinitialisation de la table interet
}
