% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{notification}
\alias{notification}
\title{Création d'une notification pour l'utilisateur}
\usage{
notification(int)
}
\arguments{
\item{int}{table propriétaire avec uniquement les lignes les parcelles qui intéressent le destinataire et dont le propriétaire à modifié la destination.}
}
\value{
renseigne db notification
}
\description{
Création d'une notification pour l'utilisateur
}
