% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{r_input}
\alias{r_input}
\title{Intègre un input à l'objet reactif r}
\usage{
r_input(id, r, input, output, session)
}
\arguments{
\item{id}{nom input}

\item{r}{reactive value}

\item{input}{input}

\item{output}{output}

\item{session}{session}
}
\value{
r
}
\description{
Intègre un input à l'objet reactif r
}
