% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{read_identite}
\alias{read_identite}
\title{lecture base de donnée identite}
\usage{
read_identite(inv = FALSE)
}
\arguments{
\item{inv}{si TRUE, retourne aussi les inscriptions non validées}
}
\value{
data frame
}
\description{
lecture base de donnée identite
}
