% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{setup_parcelles}
\alias{setup_parcelles}
\title{Création du fichier sf des parcelles}
\usage{
setup_parcelles(path)
}
\arguments{
\item{path}{chemin du dossier}
}
\value{
rien : écrit parcelles.rds
}
\description{
Création du fichier sf des parcelles
}
