% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{update_data}
\alias{update_data}
\title{mise à jour de r$data}
\usage{
update_data(r, input, output, session, ini = FALSE)
}
\arguments{
\item{r}{r}

\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{ini}{TRUE pour initialiser r$data}
}
\value{
r
}
\description{
mise à jour de r$data
}
