% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{write_interet}
\alias{write_interet}
\title{enregistre base de donnée interet}
\usage{
write_interet(pr)
}
\arguments{
\item{pr}{table}
}
\value{
data frame
}
\description{
enregistre base de donnée interet
}
