% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{write_proprietaire}
\alias{write_proprietaire}
\title{enregistre base de donnée proprietaire}
\usage{
write_proprietaire(r)
}
\arguments{
\item{pr}{table}
}
\value{
rien
}
\description{
enregistre base de donnée proprietaire
}
