% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{zoom_parcelle}
\alias{zoom_parcelle}
\title{Zoom sur parcelle sélectionnée}
\usage{
zoom_parcelle(input, output, session, r)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{r}{r}
}
\value{
coordonnées
}
\description{
Zoom sur parcelle sélectionnée
}
